+++ quilt-sync/src-tauri/src/telemetry/tracing.rs
@@ -19,6 +19,11 @@ impl LogsDir {
             LogsDir::Temporary(temp_dir) => temp_dir.path(),
         }
     }
+
+    #[cfg(test)]
+    pub fn create() -> Result<Self> {
+        Ok(LogsDir::Temporary(tempfile::TempDir::new()?))
+    }
 }
 
 fn get_logs_dir(base_path: &Path) -> Result<LogsDir> {